mqttazureiotedge/MQTT Controller Code/requirements.txt
azure-data-tables
paho-mqtt
azure~=4.0.0
termcolor~=1.1.0